Feat: Added support for custom express endpoints from configuration
