+++ scripts/server.js
@@ -1,27 +1,12 @@
 'use strict';
 
-const express = require('express');
-const webpack = require('webpack');
-const makeWebpackConfig = require('./make-webpack-config');
+const createServer = require('./create-server');
 
 module.exports = function server(config, callback) {
-	const app = express();
-	const webpackConfig = makeWebpackConfig(config, 'development');
-	const stats = webpackConfig.stats || {};
-	const compiler = webpack(webpackConfig);
+	const env = 'development';
+	const serverInfo = createServer(config, env);
 
-	app.use(require('webpack-dev-middleware')(compiler, {
-		noInfo: true,
-		stats,
-	}));
+	serverInfo.app.listen(config.serverPort, config.serverHost, callback);
 
-	app.use(require('webpack-hot-middleware')(compiler));
-
-	if (config.assetsDir) {
-		app.use(express.static(config.assetsDir));
-	}
-
-	app.listen(config.serverPort, config.serverHost, callback);
-
-	return compiler;
+	return serverInfo.compiler;
 };
